pyproject.toml
[project]
name = "cl_ml_tools"
version = "0.2.1"
description = "Tools for master-worker media processing / ML - provides task plugins, worker runtime, and protocols for job persistence and file storage"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Ananda Sarangaram"}
]
keywords = ["media", "ml", "processing", ]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pydantic>=2.0",
    "paho-mqtt>=1.6.0",
    "fastapi>=0.100",
    "python-multipart>=0.0.6",
    "Pillow>=10.0",
    "ruff>=0.14.9",
    "opencv-python-headless>=4.12.0.88",
    "basedpyright>=1.36.1",
    "python-magic>=0.4.27",
    "pillow-heif>=1.1.1",
    "m3u8>=6.0.0",
    "numpy>=1.24.0",
    "onnx>=1.15.0",
    "onnxruntime>=1.16.0",
    "httpx>=0.24",
    "scipy>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "httpx>=0.24",
]
all = [
    "cl_ml_tools[dev]",
]

[project.entry-points."cl_ml_tools.tasks"]
media_thumbnail = "cl_ml_tools.plugins.media_thumbnail.task:MediaThumbnailTask"
image_conversion = "cl_ml_tools.plugins.image_conversion.task:ImageConversionTask"
hash = "cl_ml_tools.plugins.hash.task:HashTask"
hls_streaming = "cl_ml_tools.plugins.hls_streaming.task:HLSStreamingTask"
exif = "cl_ml_tools.plugins.exif.task:ExifTask"
face_detection = "cl_ml_tools.plugins.face_detection.task:FaceDetectionTask"
face_embedding = "cl_ml_tools.plugins.face_embedding.task:FaceEmbeddingTask"
dino_embedding = "cl_ml_tools.plugins.dino_embedding.task:DinoEmbeddingTask"
clip_embedding = "cl_ml_tools.plugins.clip_embedding.task:ClipEmbeddingTask"

[project.entry-points."cl_ml_tools.routes"]
media_thumbnail = "cl_ml_tools.plugins.media_thumbnail.routes:create_router"
image_conversion = "cl_ml_tools.plugins.image_conversion.routes:create_router"
hash = "cl_ml_tools.plugins.hash.routes:create_router"
hls_streaming = "cl_ml_tools.plugins.hls_streaming.routes:create_router"
exif = "cl_ml_tools.plugins.exif.routes:create_router"
face_detection = "cl_ml_tools.plugins.face_detection.routes:create_router"
face_embedding = "cl_ml_tools.plugins.face_embedding.routes:create_router"
dino_embedding = "cl_ml_tools.plugins.dino_embedding.routes:create_router"
clip_embedding = "cl_ml_tools.plugins.clip_embedding.routes:create_router"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cl_ml_tools"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
addopts = "--cov=src/cl_ml_tools --cov-report=html --cov-report=term-missing --cov-fail-under=85"

[tool.coverage.run]
source = ["src/cl_ml_tools"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
precision = 2
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
